Reading and attaching of data to response object on server side up
Now we just need to send it over the wire. Also there's a strange issue
where this will work for the first few requests, and then stop working
(I think, the print statements would suggest this). Probably has
something to do with closing sockets or something like that, next on the
list